completed LifelineConnection ResidentialConnection

+++ Lab-08/GridLedger/src/main/java/Connection.java
@@ -7,7 +7,7 @@ public class Connection {
 
     }
 
-    public Byte energyCharge() {
+    public double energyCharge() {
         return 0;
     }
